fix: Fix roomChattingUI
