+++ Assets/Scripts/JBB/LobbySceneManager.cs
@@ -88,7 +88,8 @@ namespace JBB
             PhotonNetwork.LocalPlayer.SetNickname(GameManager.Chat.Nickname);
             PhotonNetwork.LocalPlayer.SetReady(false);
             PhotonNetwork.LocalPlayer.SetLoad(false);
-            GameManager.Scene.LoadScene("RoomScene");
+            PhotonNetwork.LoadLevel("RoomScene");
+            //GameManager.Scene.LoadScene("RoomScene");
         }
         public override void OnJoinRoomFailed(short returnCode, string message)
         {
