+++ Assets/Scripts/JBB/RoomChattingPrefab.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 091d8c49a5cb3f441a9d05eaa37c5dca
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
